+++ HISTORY.rst
@@ -4,6 +4,10 @@ Version History
 v0.8.0 (unreleased)
 -------------------
 
+New Features
+^^^^^^^^^^^^
+* ``clisops.core.average.average_shape`` copies the global and variable attributes from the input data to the results.
+
 Bug fixes
 ---------
 * ``average_shape`` and ``create_weight_masks`` were adapted to work with xESMF 0.6.1, while maintaining compatibility with earlier versions.